Move util function extend into its own file.
