+++ lib/wsdl/index.js
@@ -21,6 +21,7 @@ var stripBom = require('strip-bom');
 
 var HttpClient = require('../http');
 var splitNSName = require('./split_ns_name');
+var extend = require('./extend');
 
 var Primitives = {
   string: 1,
@@ -75,25 +76,6 @@ var trimRight = /[\s\xA0]+$/;
 
 function trim(text) {
   return text.replace(trimLeft, '').replace(trimRight, '');
-}
-
-/**
- * What we want is to copy properties from one object to another one and avoid 
- * properties overriding. This way we ensure the 'inheritance' of 
- * <xsd:extension base=...> usage.
- *
- * NB: 'Element' (and subtypes) don't have any prototyped properties: there's 
- * no need to process a 'hasOwnProperties' call, we should just iterate over the 
- * keys.
- */
-function extend(base, obj) {
-  if(base !== null && typeof base === "object" && obj !== null && typeof obj === "object"){
-    Object.keys(obj).forEach(function(key) {
-      if(!base.hasOwnProperty(key))
-        base[key] = obj[key];
-    });
-  }
-  return base;
 }
 
 function deepMerge(destination, source) {
